requirements.txt
numpy==1.22.0
pandas==1.3.5
Pillow==9.0.0
pygame==2.1.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
